embedded-biometric-approval-system/backend/requirements.txt
 # Flask and Extensions
flask==2.1.0
flask-sqlalchemy==2.5.1
flask-cors==3.0.10

# Database
psycopg2-binary==2.9.3
sqlalchemy==1.4.32

# Authentication and Security
pyjwt==2.3.0
werkzeug==2.0.3
cryptography==3.4.8

# Blockchain
web3==5.27.0

# Utilities
python-dotenv==0.19.2

# Deployment
gunicorn==20.1.0

# Development and Testing
pytest==7.0.1
coverage==6.3.1

# Logging
python-json-logger==2.0.2
